Music reworked. Allow play music by query instead of URL; Music can be played across multiple servers (different queues).

+++ src/app/music/playlists/adapter.js
@@ -7,8 +7,8 @@ const add = require('./helpers/addToPlaylist');
 const remove = require('./helpers/removeFromPlaylist');
 
 module.exports = {
-    play: function(message) {play(message)},
-    list: function(message) {list(message)},
+    play: function(message, playlistName, doShuffle) {play(message, playlistName, doShuffle)},
+    list: function(message, playlistName, includeLinks) {list(message, playlistName, includeLinks)},
     listAll: function(message) {
         listDir(paths.getPlaylistPath(message))
         .then(list => {
@@ -23,20 +23,32 @@ module.exports = {
         })
         .catch(e => console.log(e));
     },
-    create: function(message) {create(message)},
-    add: async function(message) {
-        let result = await add(message, paths.getPlaylistPath(message));
-        console.log((result) ? 'Added song successfully!' : 'Failed to add song!');
-        message.channel.send((result) ? `Added song to the playlist!\n${message.content.split(' ')[3]}` : 'Failed to add song to playlist');
-        try {
-            message.delete();
-        }
-        catch (e) {
-            message.channel.send('I require admin permissions to operate correctly.');
-        }
+    create: function(message, playlistName) {create(message, playlistName)},
+    add: async function(message, playlistName, songURL) {
+        add(message, playlistName, songURL)
+        .then(result => {
+            if (result !== undefined) {
+                message.channel.send((result) ? `Added song to the playlist!\n${message.content.split(' ')[3]}` : 'Failed to add song to playlist');
+                try {
+                    message.delete();
+                }
+                catch (e) {
+                    message.channel.send('I require admin permissions to operate correctly.');
+                }
+            }
+        })
+        .catch(e => {
+            message.channel.send(e.message || e)
+            try {
+                message.delete();
+            }
+            catch (e) {
+                message.channel.send('I require admin permissions to operate correctly.');
+            }
+        });
     },
-    remove: async function(message) {
-        let result = await remove(message, paths.getPlaylistPath(message));
+    remove: async function(message, playlistName, index) {
+        let result = await remove(message, playlistName, index);
         console.log((result) ? 'Removed song successfully!' : 'Failed to remove song!');
         message.channel.send((result) ? 'Removed song from the playlist!' : 'Failed to remove song from playlist');
         try {
